// src/shared/api/models/getApiGiftCardsCodeCode200.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiGiftCardsCodeCode200GiftCardCode } from './getApiGiftCardsCodeCode200GiftCardCode';
import type { GetApiGiftCardsCodeCode200RecipientName } from './getApiGiftCardsCodeCode200RecipientName';
import type { GetApiGiftCardsCodeCode200Status } from './getApiGiftCardsCodeCode200Status';
import type { GetApiGiftCardsCodeCode200ExpiryDate } from './getApiGiftCardsCodeCode200ExpiryDate';
import type { GetApiGiftCardsCodeCode200PersonalMessage } from './getApiGiftCardsCodeCode200PersonalMessage';
import type { GetApiGiftCardsCodeCode200SoldBy } from './getApiGiftCardsCodeCode200SoldBy';
import type { GetApiGiftCardsCodeCode200Note } from './getApiGiftCardsCodeCode200Note';
import type { GetApiGiftCardsCodeCode200CustomerId } from './getApiGiftCardsCodeCode200CustomerId';
import type { GetApiGiftCardsCodeCode200GiftCardOnlineId } from './getApiGiftCardsCodeCode200GiftCardOnlineId';
import type { GetApiGiftCardsCodeCode200Customer } from './getApiGiftCardsCodeCode200Customer';
import type { GetApiGiftCardsCodeCode200HistoryItem } from './getApiGiftCardsCodeCode200HistoryItem';

export type GetApiGiftCardsCodeCode200 = {
    id: string;
    /**
     * @minimum -2147483648
     * @maximum 2147483647
     */
    seqId: number;
    /**
     * @minimum -2147483648
     * @maximum 2147483647
     */
    outletId: number;
    giftCardCode: GetApiGiftCardsCodeCode200GiftCardCode;
    /**
     * @minimum -140737488355328
     * @maximum 140737488355327
     */
    originalValue: number;
    /**
     * @minimum -140737488355328
     * @maximum 140737488355327
     */
    residueValue: number;
    recipientName: GetApiGiftCardsCodeCode200RecipientName;
    status: GetApiGiftCardsCodeCode200Status;
    applicableServiceIds?: string[];
    expiryDate: GetApiGiftCardsCodeCode200ExpiryDate;
    taxIds?: string[];
    /**
     * @minimum -140737488355328
     * @maximum 140737488355327
     */
    taxAmount: number;
    personalMessage: GetApiGiftCardsCodeCode200PersonalMessage;
    soldBy: GetApiGiftCardsCodeCode200SoldBy;
    createdAt: string;
    note: GetApiGiftCardsCodeCode200Note;
    /**
     * @minimum -2147483648
     * @maximum 2147483647
     */
    sourceId: number;
    customerId: GetApiGiftCardsCodeCode200CustomerId;
    giftCardOnlineId: GetApiGiftCardsCodeCode200GiftCardOnlineId;
    updatedAt: string;
    customer?: GetApiGiftCardsCodeCode200Customer;
    history?: GetApiGiftCardsCodeCode200HistoryItem[];
};